Update config.jelly
Fix for https://issues.jenkins-ci.org/browse/JENKINS-63055

We found that the cause of the issue appears to be the use of an additional jelly tag (in this case f:description) within the drop down f:entry tag.  This prevented the creation of the <td class=“setting-no-help”></td> tag which in turn caused the drop down refresh to fail (as the setting-main tag now requires a sibling).

Simply moving the f:description jelly tag outside the f:entry tag fixed the issue.